backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.14.1
reportlab==4.2.5
slowapi==0.1.9
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0
email-validator==2.2.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
python-multipart==0.0.6
